% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_icd_selector.R
\name{mod_icd_selector_server}
\alias{mod_icd_selector_server}
\title{ICD Selector Module Server}
\usage{
mod_icd_selector_server(id)
}
\arguments{
\item{id}{Module namespace id}
}
\value{
Reactive expression returning vector of selected ICD codes
}
\description{
Server logic for ICD code selector. Loads ICD data and updates
virtualSelectInput choices.
}
