% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_icd_selector.R
\name{mod_icd_selector_ui}
\alias{mod_icd_selector_ui}
\title{ICD Selector Module UI}
\usage{
mod_icd_selector_ui(id)
}
\arguments{
\item{id}{Module namespace id}
}
\value{
Shiny UI tag list
}
\description{
Creates a virtual select input for choosing ICD-10 codes.
Uses virtualSelectInput for performance with 40k+ items.
}
