% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap.R
\name{mod_treemap_server}
\alias{mod_treemap_server}
\title{Treemap Module Server}
\usage{
mod_treemap_server(id, selected_codes)
}
\arguments{
\item{id}{Module namespace id}

\item{selected_codes}{Reactive expression returning vector of selected ICD codes}
}
\value{
NULL (side effects only)
}
\description{
Treemap Module Server
}
