% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_treemap.R
\name{mod_treemap_server}
\alias{mod_treemap_server}
\title{Treemap Module Server}
\usage{
mod_treemap_server(id, r)
}
\arguments{
\item{id}{Module namespace id}

\item{r}{Shared reactiveValues object with selected_codes slot}
}
\value{
NULL (side effects only)
}
\description{
Treemap Module Server
}
